requirements.txt
mysql-connector-python
flask
python
